Translations, Fix: Frontend Article_Cell

+++ admin/Views/Partials/table_action.php
@@ -9,9 +9,20 @@
             <i class="material-icons">create</i>
         </a>
     </li>
-    <li>
-        <a href="javascript:void(0)" @click="onDeleteItem('<?= $options['controller'] ?>', <?= $options['id'] ?>)">
-            <i class="material-icons">delete</i>
-        </a>
-    </li>
+    <?php
+    if (!isset($options['is_system'])): ?>
+        <li>
+            <a href="javascript:void(0)" @click="onDeleteItem('<?= $options['controller'] ?>', <?= $options['id'] ?>)">
+                <i class="material-icons">delete</i>
+            </a>
+        </li>
+    <?php
+    elseif (!$options['is_system']): ?>
+        <li>
+            <a href="javascript:void(0)" @click="onDeleteItem('<?= $options['controller'] ?>', <?= $options['id'] ?>)">
+                <i class="material-icons">delete</i>
+            </a>
+        </li>
+    <?php
+    endif; ?>
 </ul>
